refactors and deletes unusued components
